+++ src/simulators/superoperator/superoperator_state.hpp
@@ -12,8 +12,8 @@
  * that they have been altered from the originals.
  */
 
-#ifndef _unitary_state_hpp
-#define _unitary_state_hpp
+#ifndef _superoperator_state_hpp
+#define _superoperator_state_hpp
 
 #include <algorithm>
 #define _USE_MATH_DEFINES
@@ -22,28 +22,30 @@
 #include "framework/utils.hpp"
 #include "framework/json.hpp"
 #include "base/state.hpp"
-#include "unitarymatrix.hpp"
+#include "superoperator.hpp"
 
 
 namespace AER {
-namespace QubitUnitary {
+namespace QubitSuperoperator {
 
 // Allowed gates enum class
 enum class Gates {
-  id, h, s, sdg, t, tdg, // single qubit
-  // multi-qubit controlled (including single-qubit non-controlled)
-  mcx, mcy, mcz, mcu1, mcu2, mcu3, mcswap
+  u1, u2, u3, id, x, y, z, h, s, sdg, t, tdg, // single qubit
+  cx, cz, swap, // two qubit
+  ccx // three qubit
 };
 
+// Allowed snapshots enum class
+enum class Snapshots {superop};
 
 //=========================================================================
 // QubitUnitary State subclass
 //=========================================================================
 
-template <class data_t = double>
-class State : public Base::State<QV::UnitaryMatrix<data_t>> {
+template <class data_t = QV::Superoperator<double>>
+class State : public Base::State<data_t> {
 public:
-  using BaseState = Base::State<QV::UnitaryMatrix<data_t>>;
+  using BaseState = Base::State<data_t>;
 
   State() = default;
   virtual ~State() = default;
@@ -53,28 +55,30 @@ public:
   //-----------------------------------------------------------------------
 
   // Return the string name of the State class
-  virtual std::string name() const override {return "unitary";}
+  virtual std::string name() const override {return "superoperator";}
 
   // Return the set of qobj instruction types supported by the State
   virtual Operations::OpSet::optypeset_t allowed_ops() const override {
     return Operations::OpSet::optypeset_t({
       Operations::OpType::gate,
+      Operations::OpType::reset,
+      Operations::OpType::snapshot,
       Operations::OpType::barrier,
       Operations::OpType::matrix,
-      Operations::OpType::snapshot
+      Operations::OpType::kraus,
+      Operations::OpType::superop
     });
   }
 
   // Return the set of qobj gate instruction names supported by the State
   virtual stringset_t allowed_gates() const override {
-    return {"u1", "u2", "u3", "cx", "cz", "cy", "cu1", "swap",
-            "id", "x", "y", "z", "h", "s", "sdg", "t", "tdg", "ccx",
-            "mcx", "mcz", "mcy", "mcu1", "mcu2", "mcu3", "mcswap"};
+    return {"U", "CX", "u1", "u2", "u3", "cx", "cz", "swap",
+            "id", "x", "y", "z", "h", "s", "sdg", "t", "tdg", "ccx"};
   }
 
   // Return the set of qobj snapshot types supported by the State
   virtual stringset_t allowed_snapshots() const override {
-    return {"unitary"};
+    return {"superoperator"};
   }
 
   // Apply a sequence of operations by looping over list
@@ -86,20 +90,20 @@ public:
   // Initializes an n-qubit unitary to the identity matrix
   virtual void initialize_qreg(uint_t num_qubits) override;
 
-  // Initializes to a specific n-qubit unitary matrix
+  // Initializes to a specific n-qubit unitary superop
   virtual void initialize_qreg(uint_t num_qubits,
-                               const QV::UnitaryMatrix<data_t> &unitary) override;
+                               const data_t &unitary) override;
 
   // Returns the required memory for storing an n-qubit state in megabytes.
   // For this state the memory is indepdentent of the number of ops
-  // and is approximately 16 * 1 << 2 * num_qubits bytes
+  // and is approximately 16 * 1 << 4 * num_qubits bytes
   virtual size_t required_memory_mb(uint_t num_qubits,
                                     const std::vector<Operations::Op> &ops)
                                     const override;
 
   // Load the threshold for applying OpenMP parallelization
   // if the controller/engine allows threads for it
-  // Config: {"omp_qubit_threshold": 7}
+  // Config: {"omp_qubit_threshold": 3}
   virtual void set_config(const json_t &config) override;
 
   //-----------------------------------------------------------------------
@@ -133,6 +137,15 @@ protected:
   // Apply a matrix to given qubits (identity on all other qubits)
   void apply_matrix(const reg_t &qubits, const cvector_t & vmat);
 
+  // Reset the specified qubits to the |0> state by simulating
+  // a measurement, applying a conditional x-gate if the outcome is 1, and
+  // then discarding the outcome.
+  void apply_reset(const reg_t &qubits);
+
+  // Apply a Kraus error operation
+  void apply_kraus(const reg_t &qubits,
+                   const std::vector<cmatrix_t> &krausops);
+
   //-----------------------------------------------------------------------
   // 1-Qubit Gates
   //-----------------------------------------------------------------------
@@ -147,17 +160,17 @@ protected:
   // Apply N-qubit multi-controlled single qubit waltz gate specified by
   // parameters u3(theta, phi, lambda)
   // NOTE: if N=1 this is just a regular u3 gate.
-  void apply_gate_mcu3(const reg_t& qubits,
-                       const double theta,
-                       const double phi,
-                       const double lambda);
+  void apply_gate_u3(const uint_t qubit,
+                     const double theta,
+                     const double phi,
+                     const double lambda);
 
   //-----------------------------------------------------------------------
   // Config Settings
   //-----------------------------------------------------------------------
 
   // OpenMP qubit threshold
-  int omp_qubit_threshold_ = 6;
+  int omp_qubit_threshold_ = 3;
 
   // Threshold for chopping small values to zero in JSON
   double json_chop_threshold_ = 1e-10;
@@ -174,34 +187,27 @@ protected:
 template <class data_t>
 const stringmap_t<Gates> State<data_t>::gateset_({
   // Single qubit gates
-  {"id", Gates::id},   // Pauli-Identity gate
-  {"x", Gates::mcx},    // Pauli-X gate
-  {"y", Gates::mcy},    // Pauli-Y gate
-  {"z", Gates::mcz},    // Pauli-Z gate
-  {"s", Gates::s},    // Phase gate (aka sqrt(Z) gate)
-  {"sdg", Gates::sdg}, // Conjugate-transpose of Phase gate
-  {"h", Gates::h},    // Hadamard gate (X + Z / sqrt(2))
-  {"t", Gates::t},    // T-gate (sqrt(S))
-  {"tdg", Gates::tdg}, // Conjguate-transpose of T gate
+  {"id", Gates::id},     // Pauli-Identity gate
+  {"x", Gates::x},       // Pauli-X gate
+  {"y", Gates::y},       // Pauli-Y gate
+  {"z", Gates::z},       // Pauli-Z gate
+  {"s", Gates::s},       // Phase gate (aka sqrt(Z) gate)
+  {"sdg", Gates::sdg},   // Conjugate-transpose of Phase gate
+  {"h", Gates::h},       // Hadamard gate (X + Z / sqrt(2))
+  {"t", Gates::t},       // T-gate (sqrt(S))
+  {"tdg", Gates::tdg},   // Conjguate-transpose of T gate
   // Waltz Gates
-  {"u1", Gates::mcu1},  // zero-X90 pulse waltz gate
-  {"u2", Gates::mcu2},  // single-X90 pulse waltz gate
-  {"u3", Gates::mcu3},  // two X90 pulse waltz gate
+  {"u1", Gates::u1},     // zero-X90 pulse waltz gate
+  {"u2", Gates::u2},     // single-X90 pulse waltz gate
+  {"u3", Gates::u3},     // two X90 pulse waltz gate
+  {"U", Gates::u3},      // two X90 pulse waltz gate
   // Two-qubit gates
-  {"cx", Gates::mcx},   // Controlled-X gate (CNOT)
-  {"cy", Gates::mcy},   // Controlled-Z gate
-  {"cz", Gates::mcz},   // Controlled-Z gate
-  {"cu1", Gates::mcu1}, // Controlled-u1 gate
-  {"swap", Gates::mcswap}, // SWAP gate
+  {"CX", Gates::cx},     // Controlled-X gate (CNOT)
+  {"cx", Gates::cx},     // Controlled-X gate (CNOT)
+  {"cz", Gates::cz},     // Controlled-Z gate
+  {"swap", Gates::swap}, // SWAP gate
   // Three-qubit gates
-  {"ccx", Gates::mcx},   // Controlled-CX gate (Toffoli)
-  {"mcx", Gates::mcx},   // Multi-controlled-X gate
-  {"mcy", Gates::mcy},   // Multi-controlled-Y gate
-  {"mcz", Gates::mcz},    // Multi-controlled-Z gate
-  {"mcu1", Gates::mcu1}, // Multi-controlled-u1
-  {"mcu2", Gates::mcu2}, // Multi-controlled-u2
-  {"mcu3", Gates::mcu3},  // Multi-controlled-u3
-  {"mcswap", Gates::mcswap} // Multi-controlled-SWAP gate
+  {"ccx", Gates::ccx}    // Controlled-CX gate (Toffoli)
 });
 
 //============================================================================
@@ -222,14 +228,23 @@ void State<data_t>::apply_ops(const std::vector<Operations::Op> &ops,
         if (BaseState::creg_.check_conditional(op))
           apply_gate(op);
         break;
-      case Operations::OpType::snapshot:
-        apply_snapshot(op, data);
+      case Operations::OpType::reset:
+        apply_reset(op.qubits);
         break;
       case Operations::OpType::matrix:
         apply_matrix(op.qubits, op.mats[0]);
         break;
+      case Operations::OpType::kraus:
+        apply_kraus(op.qubits, op.mats);
+        break;
+      case Operations::OpType::superop:
+        BaseState::qreg_.apply_superop_matrix(op.qubits, Utils::vectorize_matrix(op.mats[0]));
+        break;
+      case Operations::OpType::snapshot:
+        apply_snapshot(op, data);
+        break;
       default:
-        throw std::invalid_argument("QubitUnitary::State::invalid instruction \'" +
+        throw std::invalid_argument("QubitSuperoperator::State::invalid instruction \'" +
                                     op.name + "\'.");
     }
   }
@@ -239,11 +254,11 @@ template <class data_t>
 size_t State<data_t>::required_memory_mb(uint_t num_qubits,
                                  const std::vector<Operations::Op> &ops)
                                  const {
-  // An n-qubit unitary as 2^2n complex doubles
+  // An n-qubit unitary as 2^4n complex doubles
   // where each complex double is 16 bytes
   (void)ops; // avoid unused variable compiler warning
   size_t shift_mb = std::max<int_t>(0, num_qubits + 4 - 20);
-  size_t mem_mb = 1ULL << (2 * shift_mb);
+  size_t mem_mb = 1ULL << (4 * shift_mb);
   return mem_mb;
 }
 
@@ -251,7 +266,7 @@ size_t State<data_t>::required_memory_mb(uint_t num_qubits,
 template <class data_t>
 void State<data_t>::set_config(const json_t &config) {
   // Set OMP threshold for state update functions
-  JSON::get_value(omp_qubit_threshold_, "unitary_parallel_threshold", config);
+  JSON::get_value(omp_qubit_threshold_, "superoperator_parallel_threshold", config);
 
   // Set threshold for truncating snapshots
   JSON::get_value(json_chop_threshold_, "zero_threshold", config);
@@ -269,28 +284,32 @@ void State<data_t>::initialize_qreg(uint_t num_qubits) {
 
 template <class data_t>
 void State<data_t>::initialize_qreg(uint_t num_qubits,
-                                    const QV::UnitaryMatrix<data_t> &unitary) {
+                                    const data_t &supermat) {
   // Check dimension of state
-  if (unitary.num_qubits() != num_qubits) {
-    throw std::invalid_argument("Unitary::State::initialize: initial state does not match qubit number");
+  if (supermat.num_qubits() != num_qubits) {
+    throw std::invalid_argument("QubitSuperoperator::State::initialize: initial state does not match qubit number");
   }
   initialize_omp();
   BaseState::qreg_.set_num_qubits(num_qubits);
   const size_t sz = 1ULL << BaseState::qreg_.size();
-  BaseState::qreg_.initialize_from_data(unitary.data(), sz);
+  BaseState::qreg_.initialize_from_data(supermat.data(), sz);
 }
 
 
 template <class data_t>
 void State<data_t>::initialize_qreg(uint_t num_qubits,
-                                    const cmatrix_t &unitary) {
+                                    const cmatrix_t &mat) {
+  
   // Check dimension of unitary
-  if (unitary.size() != 1ULL << (2 * num_qubits)) {
-    throw std::invalid_argument("Unitary::State::initialize: initial state does not match qubit number");
+  const auto sz_uni = 1ULL << (2 * num_qubits);
+  const auto sz_super = 1ULL << (4 * num_qubits);
+  if (mat.size() != sz_uni && mat.size() != sz_super) {
+    throw std::invalid_argument(
+      "QubitSuperoperator::State::initialize: initial state does not match qubit number");
   }
   initialize_omp();
   BaseState::qreg_.set_num_qubits(num_qubits);
-  BaseState::qreg_.initialize_from_matrix(unitary);
+  BaseState::qreg_.initialize_from_matrix(mat);
 }
 
 
@@ -301,6 +320,34 @@ void State<data_t>::initialize_omp() {
     BaseState::qreg_.set_omp_threads(BaseState::threads_); // set allowed OMP threads in qubitvector
 }
 
+//=========================================================================
+// Implementation: Reset
+//=========================================================================
+
+template <class data_t>
+void State<data_t>::apply_reset(const reg_t &qubits) {
+  // TODO: This can be more efficient by adding reset
+  // to base class rather than doing a matrix multiplication
+  // where all but 1 row is zeros.
+  const auto reset_op = Utils::SMatrix::reset(1ULL << qubits.size());
+  BaseState::qreg_.apply_superop_matrix(qubits, Utils::vectorize_matrix(reset_op));
+}
+
+//=========================================================================
+// Implementation: Kraus Noise
+//=========================================================================
+
+template <class statevec_t>
+void State<statevec_t>::apply_kraus(const reg_t &qubits,
+                                    const std::vector<cmatrix_t> &kmats) {
+  // Convert to Superoperator
+  const auto nrows = kmats[0].GetRows();
+  cmatrix_t superop(nrows * nrows, nrows * nrows);
+  for (const auto kraus : kmats) {
+    superop += Utils::tensor_product(Utils::conjugate(kraus), kraus);
+  }
+  BaseState::qreg_.apply_superop_matrix(qubits, Utils::vectorize_matrix(superop));
+}
 
 //=========================================================================
 // Implementation: Gates
@@ -313,24 +360,41 @@ void State<data_t>::apply_gate(const Operations::Op &op) {
   if (it == gateset_.end())
     throw std::invalid_argument("Unitary::State::invalid gate instruction \'" +
                                 op.name + "\'.");
-  Gates g = it -> second;
-  switch (g) {
-    case Gates::mcx:
-      // Includes X, CX, CCX, etc
-      BaseState::qreg_.apply_mcx(op.qubits);
+  switch (it -> second) {
+    case Gates::u3:
+      apply_gate_u3(op.qubits[0],
+                    std::real(op.params[0]),
+                    std::real(op.params[1]),
+                    std::real(op.params[2]));
       break;
-    case Gates::mcy:
-      // Includes Y, CY, CCY, etc
-      BaseState::qreg_.apply_mcy(op.qubits);
+    case Gates::u2:
+      apply_gate_u3(op.qubits[0],
+                    M_PI / 2.,
+                    std::real(op.params[0]),
+                    std::real(op.params[1]));
       break;
-    case Gates::mcz:
-      // Includes Z, CZ, CCZ, etc
-      BaseState::qreg_.apply_mcphase(op.qubits, -1);
+    case Gates::u1:
+      apply_gate_phase(op.qubits[0], std::exp(complex_t(0., 1.) * op.params[0]));
+      break;
+    case Gates::cx:
+      BaseState::qreg_.apply_cnot(op.qubits[0], op.qubits[1]);
+      break;
+    case Gates::cz:
+      BaseState::qreg_.apply_cz(op.qubits[0], op.qubits[1]);
       break;
     case Gates::id:
       break;
+    case Gates::x:
+      BaseState::qreg_.apply_x(op.qubits[0]);
+      break;
+    case Gates::y:
+      BaseState::qreg_.apply_y(op.qubits[0]);
+      break;
+    case Gates::z:
+      BaseState::qreg_.apply_z(op.qubits[0]);
+      break;
     case Gates::h:
-      apply_gate_mcu3(op.qubits, M_PI / 2., 0., M_PI);
+      apply_gate_u3(op.qubits[0], M_PI / 2., 0., M_PI);
       break;
     case Gates::s:
       apply_gate_phase(op.qubits[0], complex_t(0., 1.));
@@ -346,31 +410,15 @@ void State<data_t>::apply_gate(const Operations::Op &op) {
       const double isqrt2{1. / std::sqrt(2)};
       apply_gate_phase(op.qubits[0], complex_t(isqrt2, -isqrt2));
     } break;
-    case Gates::mcswap:
-      // Includes SWAP, CSWAP, etc
-      BaseState::qreg_.apply_mcswap(op.qubits);
-      break;
-    case Gates::mcu3:
-      // Includes u3, cu3, etc
-      apply_gate_mcu3(op.qubits,
-                      std::real(op.params[0]),
-                      std::real(op.params[1]),
-                      std::real(op.params[2]));
-      break;
-    case Gates::mcu2:
-      // Includes u2, cu2, etc
-      apply_gate_mcu3(op.qubits,
-                      M_PI / 2.,
-                      std::real(op.params[0]),
-                      std::real(op.params[1]));
-      break;
-    case Gates::mcu1:
-      // Includes u1, cu1, etc
-      BaseState::qreg_.apply_mcphase(op.qubits, std::exp(complex_t(0, 1) * op.params[0]));
+    case Gates::swap: {
+      BaseState::qreg_.apply_swap(op.qubits[0], op.qubits[1]);
+    } break;
+    case Gates::ccx:
+      BaseState::qreg_.apply_toffoli(op.qubits[0], op.qubits[1], op.qubits[2]);
       break;
     default:
       // We shouldn't reach here unless there is a bug in gateset
-      throw std::invalid_argument("Unitary::State::invalid gate instruction \'" +
+      throw std::invalid_argument("Superoperator::State::invalid gate instruction \'" +
                                   op.name + "\'.");
   }
 }
@@ -378,7 +426,7 @@ void State<data_t>::apply_gate(const Operations::Op &op) {
 template <class data_t>
 void State<data_t>::apply_matrix(const reg_t &qubits, const cmatrix_t &mat) {
   if (qubits.empty() == false && mat.size() > 0) {
-    apply_matrix(qubits, Utils::vectorize_matrix(mat));
+    BaseState::qreg_.apply_unitary_matrix(qubits, Utils::vectorize_matrix(mat));
   }
 }
 
@@ -386,29 +434,29 @@ template <class data_t>
 void State<data_t>::apply_matrix(const reg_t &qubits, const cvector_t &vmat) {
   // Check if diagonal matrix
   if (vmat.size() == 1ULL << qubits.size()) {
-    BaseState::qreg_.apply_diagonal_matrix(qubits, vmat);
+    BaseState::qreg_.apply_diagonal_unitary_matrix(qubits, vmat);
   } else {
-    BaseState::qreg_.apply_matrix(qubits, vmat);
+    BaseState::qreg_.apply_unitary_matrix(qubits, vmat);
   }
 }
 
 
 template <class data_t>
 void State<data_t>::apply_gate_phase(uint_t qubit, complex_t phase) {
-  cmatrix_t diag(1, 2);
-  diag(0, 0) = 1.0;
-  diag(0, 1) = phase;
-  apply_matrix(reg_t({qubit}), diag);
+  cvector_t diag(2);
+  diag[0] = 1.0;
+  diag[1] = phase;
+  BaseState::qreg_.apply_diagonal_unitary_matrix(reg_t({qubit}), diag);
 }
 
 
 template <class statevec_t>
-void State<statevec_t>::apply_gate_mcu3(const reg_t& qubits,
-                                        double theta,
-                                        double phi,
-                                        double lambda) {
-  const auto u3 = Utils::Matrix::u3(theta, phi, lambda);
-  BaseState::qreg_.apply_mcu(qubits, Utils::vectorize_matrix(u3));
+void State<statevec_t>::apply_gate_u3(const uint_t qubit,
+                                      double theta,
+                                      double phi,
+                                      double lambda) {
+  const auto u3 = Utils::VMatrix::u3(theta, phi, lambda);
+  BaseState::qreg_.apply_unitary_matrix(reg_t({qubit}), u3);
 }
 
 
@@ -416,16 +464,16 @@ template <class data_t>
 void State<data_t>::apply_snapshot(const Operations::Op &op,
                                    OutputData &data) {
   // Look for snapshot type in snapshotset
-  if (op.name == "unitary" || op.name == "state") {
-    BaseState::snapshot_state(op, data);
+  if (op.name == "superopertor" || op.name == "state") {
+    BaseState::snapshot_state(op, data, "superoperator");
   } else {
-    throw std::invalid_argument("Unitary::State::invalid snapshot instruction \'" +
+    throw std::invalid_argument("QubitSuperoperator::State::invalid snapshot instruction \'" +
                                 op.name + "\'.");
   }
 }
 
 //------------------------------------------------------------------------------
-} // end namespace QubitUnitary::
+} // end namespace QubitSuperoperator
 } // end namespace AER
 //------------------------------------------------------------------------------
 #endif