+++ test/terra/backends/qasm_simulator/matrix_product_state_method.py
@@ -10,7 +10,7 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 """
-Tensor Network Integration Tests
+Matrix product state integration tests
 This set of tests runs all the circuits from test/terra/references/: ref_1q_clifford and ref_2q_clifford,
 with the exception of those with the multiplexer gate which is not supported yet.
 """
@@ -28,12 +28,13 @@ from qiskit.providers.aer import QasmSimulator
 
 logger = logging.getLogger(__name__)
 
-class TestQasmTensorNetworkSimulator(common.QiskitAerTestCase):
-    """QasmSimulator tensor_network method tests."""
+class QasmMatrixProductStateMethodTests:
+    """QasmSimulator matrix_product_state method tests."""
 
-
-    def test_measure_deterministic_without_sampling(self):
-        """TestTensorNetwork measure with deterministic counts without sampling"""
+    BACKEND_OPTS = {"method": "matrix_product_state"}
+    
+    def test_method_deterministic_without_sampling(self):
+        """Test matrix product state method with deterministic counts without sampling"""
         deterministic_tests = [(ref_1q_clifford.h_gate_circuits_deterministic, ref_1q_clifford.h_gate_counts_deterministic),
              (ref_1q_clifford.x_gate_circuits_deterministic, ref_1q_clifford.x_gate_counts_deterministic),
              (ref_1q_clifford.z_gate_circuits_deterministic, ref_1q_clifford.z_gate_counts_deterministic),
@@ -55,7 +56,6 @@ class TestQasmTensorNetworkSimulator(common.QiskitAerTestCase):
 
         nondeterm_test_list = {'list':nondeterministic_tests, 'shots':2000, 'delta':0.05}
 
-        BACKEND_OPTS_TN = {"method": "tensor_network"}
         test_list = [determ_test_list, nondeterm_test_list]
         for list in test_list:
             for test in list['list']:
@@ -64,10 +64,8 @@ class TestQasmTensorNetworkSimulator(common.QiskitAerTestCase):
                 
                 circuits = test[0](final_measure=True)
                 targets  = test[1](shots)
-                job = execute(circuits, QasmSimulator(), backend_options=BACKEND_OPTS_TN, shots=shots)
+                job = execute(circuits, QasmSimulator(), backend_options=self.BACKEND_OPTS, shots=shots)
                 result = job.result()
                 self.is_completed(result)
                 self.compare_counts(result, circuits, targets, delta = delta*shots)
 
-if __name__ == '__main__':
-    unittest.main()