+++ CHANGELOG.md
@@ -24,12 +24,16 @@ Added
 -   Added superop qobj instruction and superoperator matrix utils (\# 289)
 -   Added support for conditional unitary, kraus, superop qobj instructions (\# 291)
 -   Add "validation_threshold" config parameter to Aer backends (\# 290)
+-   Added support for apply_measure in tensor_network_state. Also changed 
+    sample_measure to use apply_measure (\#299)
+-   Added density matrix simulation method to QasmSimulator (\# 295, \# 253)
 
-- Noise model inserter module
+
+- Noise model inserter module (\# 239)
 
 Changed
 -------
-
+-   Added density matrix method to automatic QasmSimulator methods (\# 316)
 
 Removed
 -------
@@ -40,6 +44,8 @@ Fixed
 - Bug in handling parallelization in matrix_product_state.cpp (PR \#292)
 
 - Added support for multiplication by coeff in tensor_network_state expectation value snapshots (PR \#294)
+
+- Change name of qasm simulation method from tensor_network to matrix_product_state (\#320)
 
 
 [0.2.3](https://github.com/Qiskit/qiskit-aer/compare/0.2.2...0.2.3) - 2019-07-11